requirements.txt
python-dotenv
streamlit
PyPDF2
pandas
openpyxl
spacy
transformers
torch
requests
google-generativeai